% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_web.R
\name{allsteps}
\alias{allsteps}
\title{Downloads and extract data for one specific month}
\usage{
allsteps(
  m,
  returnvalue = TRUE,
  createdatabase = FALSE,
  createrdafiles = FALSE,
  directory = tempdir(),
  varlist = varlist,
  cps.dbname = "cps.db",
  col_types = col_types
)
}
\arguments{
\item{m}{a character string in the Date format %Y%m}

\item{returnvalue}{a boolean}

\item{createrdafiles}{a boolean}

\item{directory}{a character string (Default tempdir())}

\item{varlist}{a character stringvector}

\item{cps.dbname}{(Default "cps.db",}

\item{coltypes}{a named character string vector,}
}
\description{
Downloads and extract data for one specific month
}
