% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_web.R
\name{dat.file.name.f}
\alias{dat.file.name.f}
\title{CPS data table name function}
\usage{
dat.file.name.f(m)
}
\arguments{
\item{m}{a date}
}
\description{
CPS data table name function
}
\details{
construct a table name as a function of month
}
\examples{
dat.file.name.f(as.Date("20050101","\%Y\%m\%d"))
}
