% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_web.R
\name{download.unzip.cpsmicrodata}
\alias{download.unzip.cpsmicrodata}
\title{Download and unzip cps microdata file for month m function}
\usage{
download.unzip.cpsmicrodata(m)
}
\arguments{
\item{m}{a date}
}
\value{
a character string, the path to the unzipped file
}
\description{
Download and unzip cps microdata file for month m function
}
\details{
download a zip file and unzip it in a temporary directory
}
\examples{
download.unzip.cpsmicrodata(as.Date("20050101","\%Y\%m\%d"))
}
