% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_web.R
\docType{data}
\name{format.change.dates}
\alias{format.change.dates}
\title{List of format change dates.}
\format{An object of class \code{Date} of length 13.}
\usage{
format.change.dates
}
\description{
List of format change dates.
}
\keyword{datasets}
