% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_web.R
\name{get_data_from_web}
\alias{get_data_from_web}
\title{downloads CPS data from web}
\usage{
get_data_from_web(
  startingdate = as.Date("20050101", "\%Y\%m\%d"),
  enddate = startingdate,
  varlist = tolower(c("pulineno", "pwsswgt", "hrhhid", "hrlongid", "hrhhid2", "cpsid",
    "pwcmpwgt", "hwniwgt", "PUMLR", "Pemlr", "PEHSPNON", "prtage", "peage", "PESEX",
    "GESTFIPS", "puhhmem", "prpertyp", "PTDTRACE", "prwtrace", "PRBLNONB", "GESTREC",
    "hrintsta", "hrmis")),
  col_types = list(pulineno = col_character(), pwsswgt = col_double(), hrhhid =
    col_character(), hrlongid = col_character(), hrhhid2 = col_character(), cpsid =
    col_character(), pwcmpwgt = col_double(), hwniwgt = col_double(), pumlr =
    col_character(), pemlr = col_character(), pehspnon = col_character(), prtage =
    col_double(), peage = col_integer(), pesex = col_character(), gestfips =
    col_character(), puhhmem = col_character(), prpertyp = col_character(), ptdtrace =
    col_character(), prwtrace = col_character(), prblnonb = col_character(),      gestrec
    = col_character(), hrintsta = col_character(), hrmis = col_character()),
  directory = NULL,
  createdatabase = FALSE,
  createrdafiles = FALSE,
  returnvalue = TRUE,
  cps.dbname = "cps.db"
)
}
\arguments{
\item{varlist}{a vector of character strings, the variables to keep from the downloaded file}

\item{directory}{where to save the file}

\item{createdatabase}{a boolean(default FALSE),}

\item{createrdafiles}{a boolean(default FALSE),}

\item{returnvalue}{a boolean}

\item{endgdate}{a date (default: \code{startingdate})}

\item{coltypes}{a vector of character strings, the variables to keep from the downloaded file}

\item{cps.dbname="cps.db"}{}
}
\description{
downloads CPS data from web
}
