% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_web.R
\name{instructionfilenamef}
\alias{instructionfilenamef}
\title{name of the instruction file name  as a function of month}
\usage{
instructionfilenamef(m)
}
\arguments{
\item{m}{a date}
}
\description{
name of the instruction file name  as a function of month
}
\details{
this is valid as long as the Census.gov website does not change it.
}
\examples{
instructionfilenamef(as.Date("20050101","\%Y\%m\%d"))
}
