% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_web.R
\name{last.format.change.date.f}
\alias{last.format.change.date.f}
\title{Last format change date.}
\usage{
last.format.change.date.f(m)
}
\arguments{
\item{m}{a character string}
}
\description{
Last format change date.
}
\details{
this is valid as long as the Census.gov website does not change it.
}
\examples{
last.format.change.date.f(as.Date("20050101","\%Y\%m\%d"))
}
