% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPS_utils.R
\name{monthsinperiod}
\alias{monthsinperiod}
\title{Bind monthly datasets together}
\usage{
monthsinperiod(startingmonth = "200501", endmonth = "201201")
}
\arguments{
\item{startingmonth}{a character string that is a date in the format "%Y%m" between "200501" and "201201".}

\item{endmonth}{a character string that is a date in the format "%Y%m" between "200501" and "201201".}
}
\value{
a vector
}
\description{
Bind monthly datasets together
}
\examples{
monthsinperiod("200510","200602")
}
