% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPS_utils.R
\name{rbind_period}
\alias{rbind_period}
\title{Bind monthly datasets together}
\usage{
rbind_period(startingmonth = "200501", endmonth = "201201", variables = NULL)
}
\arguments{
\item{list.tables}{a list of tables}

\item{w}{a character string: name of the  weights variable (should be the same in all tables)}

\item{list.y}{a vector of variable names}

\item{id}{a character string: name of the identifier variable  (should be the same in all tables)}

\item{groupvar}{a character string: name of the  rotation group variable (should be the same in all tables)}

\item{groups_1}{a character string:}

\item{groups0}{if \code{groupvar} is not null, a vector of possible values for L[[groupvar]]}
}
\value{

}
\description{
Bind monthly datasets together
}
\examples{
rbind_period(startingmonth="200501",endmonth="200503",c("pemlr","pwsswgt"))
}
\seealso{
CompositeRegressionEstimation::AK
}
