% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_web.R
\name{URLf}
\alias{URLf}
\title{get the URL of format file as a function of month}
\usage{
URLf(m)
}
\arguments{
\item{m}{a character string}
}
\description{
get the URL of format file as a function of month
}
\details{
this is valid as long as the Census.gov website does not change it.
}
\examples{
URLf(as.Date("20050101","\%Y\%m\%d"))
}
